Make 3D leaf render pass decision compatible with TransformInterop.
The decision about when to create a render pass at the leafward edges of
preserve-3d scenes was not compatible with the changes to preserve-3d
made in TransformInterop in a number of ways.

With TransformInterop, being the leaf of a 3D scene (where we define
leaf as the element *all* of whose descendants are flattened into it)
does not require any unusual properties on the element; it merely
requires that the parent have transform-style: preserve-3d but the
element not do so.  (Prior to TransformInterop, it required something
that created a node in the transform tree; see NeedsTransform().)

While I considered trying to create the render pass associated at the
actual leaves of the 3D scene, instead we do this at the closest
descendant that has a paint layer, which means it won't mess up stacking
context decisions.  It also likely means that in many cases we (still)
won't need to do it at all.

Doing this correctly post-TransformInterop requires that we make the
decision about whether an element is in this flattening position in the
effect node in the paint property tree while we're still operating on
layout objects and have accurate information about element parent/child
relationships.

This also avoids a case in the old code (which I'm not convinced was
bug-free in this respect prior to TransformInterop) where the code could
get set state on an effect node when operating on a transform node that
was a descendant of that effect node's transform node (but one that
shared the same effect node).

Note that the existing web test
compositing/overflow-trumps-transform-style.html is the only test that
fails if this code to cause creation of a render pass is removed.

The newly-added tests preserve3d-and-flattening-00{1,2,3}.html fail
prior to this change.   001 is the one that is closest to the broken
site; it does not fail if the code to create a render pass is just
removed.  002 and 003, however, also fail if the code is removed.  (The
distinction between 002 and 003 was needed to get more complete test
coverage with earlier versions of this change, but the revisions needed
to fix both of them mean the distinction is no longer as relevant, but I
kept them both anyway.)

Fixed: 1255544
Change-Id: I4d8ae4c6a134cc1605188779760079be1e9c7f90