[GridFragmentation] Support break-inside:avoid with MovePastBreakpoint
This patch adds support for avoiding un-appealing breaks inside items
by using MovePastBreakpoint.

If the current row for an item has container-separation, and
MovePastBreakpoint returns false - we should move the entire row into
the next fragmentainer.

This implicitly expands the intrinsic block-size of the grid.

The grid-data structures currently assume that we can't have "holes" in
our grid needed to support this. (Or at least non-uniform "holes"
grid-gap/align-items/etc are all currently uniform).

To support the grid splitting in this way it also introduces
row_offset_adjustments to accommodate pushing a particular row into the
next fragmentainer.

Bug: 614667
Change-Id: I252fecc122ade410eb9b574836d64de3d82eb72d